ingestion::common module
