+++ crates/iota-indexer/src/ingestion/common/mod.rs
@@ -1,6 +1,5 @@
 // Copyright (c) 2025 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-pub mod common;
-pub(crate) mod primary;
-pub(crate) mod snapshot;
+pub(crate) mod persist;
+pub mod prepare;
